chore(CMakeLists.txt): remove unnecessary header files from source groups fix(bstring.h): change sizeDiff variable type from ptrdiff_t to int chore(version.h): update GAME_BUILD_NUMBER to 3131

+++ src/core/bstring.h
@@ -204,7 +204,7 @@ public:
         char* found_pos = std::search(_data, _data + len(), subst, subst + subst_len);
 
         while (found_pos != _data + _len) {
-            ptrdiff_t sizeDiff = repl_len - subst_len;
+            int sizeDiff = repl_len - subst_len;
             ::memmove(found_pos + repl_len, found_pos + subst_len, _data + _len - (found_pos + subst_len) + 1);
             ::memcpy(found_pos, repl, repl_len);
             found_pos = std::search(found_pos + repl_len, _data + _len, subst, subst + subst_len);
